Move the algebras into Facet.Algebra.
